Added setup.py file and moved code out of judges_server
