+++ CMakeLists.txt
@@ -20,7 +20,7 @@ find_package(catkin REQUIRED COMPONENTS
 ## Uncomment this if the package has a setup.py. This macro ensures
 ## modules and global scripts declared therein get installed
 ## See http://ros.org/doc/api/catkin/html/user_guide/setup_dot_py.html
-# catkin_python_setup()
+catkin_python_setup()
 
 ################################################
 ## Declare ROS messages, services and actions ##
@@ -48,9 +48,8 @@ find_package(catkin REQUIRED COMPONENTS
 
 ## Generate messages in the 'msg' folder
 add_message_files(
-   FILES
-   Id.msg  
-   Position.msg  
+   FILES  
+   position.msg  
    TeamID.msg 
    TeamStatus.msg  
    Telemetry.msg  
@@ -60,6 +59,9 @@ add_message_files(
    odlc_shape.msg
    odlc_type.msg
    color.msg
+   mission.msg
+   StationaryObstacle.msg
+   FlyZone.msg
  )
 
 ## Generate services in the 'srv' folder
